refactor(pdf) Move common pdfmake code in pdfmakeutil
