+++ packages/markdown-pdf/src/TemplateMarkToPdfMakeVisitor.js
@@ -15,6 +15,10 @@
 'use strict';
 
 const { Decorators } = require('@accordproject/markdown-cicero');
+const {
+    handleFormattedText,
+    getHeadingType,
+} = require('./pdfmakeutil');
 
 /**
  * Converts a TemplateMark DOM to a PDF Make JSON.
@@ -26,75 +30,6 @@ class TemplateMarkToPdfMakeVisitor {
      * Construct the visitor
      */
     constructor() {
-    }
-
-    /**
-     * Apply marks to a leaf node
-     * @param {*} leafNode the leaf node
-     * @param {*} parameters the parameters
-     */
-    static applyMarks(leafNode, parameters) {
-        if (parameters.emph) {
-            leafNode.style = 'Emph';
-            leafNode.italics = true;
-        }
-        if (parameters.strong) {
-            leafNode.style = 'Strong';
-            leafNode.bold = true;
-        }
-        if (parameters.code) {
-            leafNode.style = 'Code';
-        }
-    }
-
-    /**
-     * Converts a formatted text node to a slate text node with marks
-     * @param {*} thing a concerto Strong, Emph or Text node
-     * @param {*} parameters the parameters
-     * @returns {*} the slate text node with marks
-     */
-    static handleFormattedText(thing, parameters) {
-        const textNode = {
-            text: TemplateMarkToPdfMakeVisitor.getText(thing)};
-
-        this.applyMarks(textNode, parameters);
-        return textNode;
-    }
-
-    /**
-     * Gets the text value from a formatted sub-tree
-     * @param {*} thing a concerto Strong, Emph or Text node
-     * @returns {string} the 'text' property of the formatted sub-tree
-     */
-    static getText(thing) {
-        if(thing.getType() === 'Text') {
-            return thing.text;
-        }
-        else {
-            if(thing.nodes && thing.nodes.length > 0) {
-                return TemplateMarkToPdfMakeVisitor.getText(thing.nodes[0]);
-            }
-            else {
-                return '';
-            }
-        }
-    }
-
-    /**
-     * Converts a heading level to a heading style name
-     * @param {*} thing concert heading node
-     * @returns {string} the heading type
-     */
-    static getHeadingType(thing) {
-        switch(thing.level) {
-        case '1': return 'heading_one';
-        case '2': return 'heading_two';
-        case '3': return 'heading_three';
-        case '4': return 'heading_four';
-        case '5': return 'heading_five';
-        case '6': return 'heading_six';
-        default: return 'heading_one';
-        }
     }
 
     /**
@@ -246,11 +181,11 @@ class TemplateMarkToPdfMakeVisitor {
         }
             break;
         case 'Text': {
-            result = TemplateMarkToPdfMakeVisitor.handleFormattedText(thing, parameters);
+            result = handleFormattedText(thing, parameters);
         }
             break;
         case 'Heading': {
-            result.style = TemplateMarkToPdfMakeVisitor.getHeadingType(thing);
+            result.style = getHeadingType(thing.level);
             result.text = this.processChildNodes(thing,parameters);
             result.tocItem = thing.nodes && thing.nodes.length > 0 ? true : false;
         }
